model_run_utils: use MAX_TOKENS as default finish_reason in generate_text_func
"OTHER" is an invalid value for caikit.interfaces.nlp.data_model.text_generation.FinishReason,
resulting in failed serialization of responses when querying the text
generation endpoint.
For `generate_text_func`, it is reasonable to assume that if the finish
reason is not `EOS_TOKEN` or `STOP_SEQUENCE`, it must be `MAX_TOKENS`.

Signed-off-by: Daniele Trifirò <dtrifiro@redhat.com>