Update sentence-transformers and allow setting trust_remote_code
sentence-transformers APIs changed. This update allows 3.x (so far).
The update is required to support some models.
Most of the update is adding new params to be in sync with the
super SentenceTransformer class. Most of these params are not used.

Support is added to allow passing the trust_remote_code paramater.
This can be done in the runtime config or with EMBEDDING_TRUST_REMOTE_CODE=true
(or =1) environment variable.

Signed-off-by: Mark Sturdevant <mark.sturdevant@ibm.com>